@@ -11,22 +11,22 @@
         }
     ],
     "require": {
-        "php": ">=5.4",
+        "php": ">=5.5",
         "twig/twig": "~1.13",
-        "symfony/framework-bundle": "^2.2|^3.0",
-        "symfony/form" : "^2.2|^3.0",
-        "symfony/validator": "^2.2|^3.0",
-        "symfony/event-dispatcher": "^2.2|^3.0",
+        "symfony/framework-bundle": "^3.0",
+        "symfony/form" : "^3.0",
+        "symfony/validator": "^3.0",
+        "symfony/event-dispatcher": "^3.0",
         "doctrine/doctrine-bundle": "^1.4",
-        "symfony/doctrine-bridge": "^2.3|^3.0",
-        "symfony/dependency-injection": "^2.6|^3.0",
-        "symfony/property-access": "^2.3|^3.0"
+        "symfony/doctrine-bridge": "^3.0",
+        "symfony/dependency-injection": "^3.0",
+        "symfony/property-access": "^3.0"
     },
     "require-dev": {
         "doctrine/orm": "~2.3",
         "phpspec/phpspec": "^2.0",
         "fsi/doctrine-extensions-bundle": "~1.0",
-        "fsi/form-extensions-bundle": "1.0.*"
+        "fsi/form-extensions-bundle": "^2.0@dev"
     },
     "suggest": {
         "fsi/doctrine-extensions-bundle": "For using fsi_file resource type",
@@ -40,7 +40,8 @@
     },
     "extra": {
         "branch-alias": {
-            "dev-master": "1.1-dev"
+            "1.1": "1.1-dev",
+            "dev-master": "2.0-dev"
         }
     }
 }
