@@ -10,7 +10,6 @@
 namespace FSi\Bundle\ResourceRepositoryBundle\Repository\Resource\Type;
 
 use Symfony\Component\Form\FormFactoryInterface;
-use Symfony\Component\HttpKernel\Kernel;
 use Symfony\Component\Validator\Constraint;
 
 abstract class AbstractType implements ResourceInterface
@@ -41,8 +40,8 @@ abstract class AbstractType implements ResourceInterface
     public function __construct($name)
     {
         $this->name = $name;
-        $this->constraints = array();
-        $this->formOptions = array();
+        $this->constraints = [];
+        $this->formOptions = [];
     }
 
     /**
@@ -104,30 +103,22 @@ abstract class AbstractType implements ResourceInterface
      */
     protected function buildFormOptions()
     {
-        $options = array(
+        $options = [
             'required' => false,
             'label' => false,
-        );
+        ];
 
         $options = array_merge($options, $this->formOptions);
 
         if (count($this->constraints)) {
             $options = array_merge(
                 $options,
-                array(
+                [
                     'constraints' => $this->constraints
-                )
+                ]
             );
         }
 
         return $options;
-    }
-
-    /**
-     * @return bool
-     */
-    protected function isFqcnRequiredAsFormType()
-    {
-        return version_compare(Kernel::VERSION, '3.0.0', '>=');
     }
 }
