@@ -25,7 +25,7 @@ class TwigFormPass implements CompilerPassInterface
 
         $container->setParameter('twig.form.resources', array_merge(
             $container->getParameter('twig.form.resources'),
-            array('@FSiResourceRepository/Form/form_div_layout.html.twig')
+            ['@FSiResourceRepository/Form/form_div_layout.html.twig']
         ));
     }
 }
